:recycle: Split ManagementApiService into smaller UseCase specific services
