+++ src/event_service.ts
@@ -1,0 +1,39 @@
+import {IIdentity} from '@essential-projects/iam_contracts';
+
+import {IConsumerApi} from '@process-engine/consumer_api_contracts';
+import {APIs, DataModels} from '@process-engine/management_api_contracts';
+
+export class EventService implements APIs.IEventManagementApi {
+
+  private readonly consumerApiService: IConsumerApi;
+
+  constructor(consumerApiService: IConsumerApi) {
+    this.consumerApiService = consumerApiService;
+  }
+
+  public async getWaitingEventsForProcessModel(identity: IIdentity, processModelId: string): Promise<DataModels.Events.EventList> {
+    return this.consumerApiService.getEventsForProcessModel(identity, processModelId);
+  }
+
+  public async getWaitingEventsForCorrelation(identity: IIdentity, correlationId: string): Promise<DataModels.Events.EventList> {
+    return this.consumerApiService.getEventsForCorrelation(identity, correlationId);
+  }
+
+  public async getWaitingEventsForProcessModelInCorrelation(
+    identity: IIdentity,
+    processModelId: string,
+    correlationId: string,
+  ): Promise<DataModels.Events.EventList> {
+
+    return this.consumerApiService.getEventsForProcessModelInCorrelation(identity, processModelId, correlationId);
+  }
+
+  public async triggerMessageEvent(identity: IIdentity, messageName: string, payload?: DataModels.Events.EventTriggerPayload): Promise<void> {
+    return this.consumerApiService.triggerMessageEvent(identity, messageName, payload);
+  }
+
+  public async triggerSignalEvent(identity: IIdentity, signalName: string, payload?: DataModels.Events.EventTriggerPayload): Promise<void> {
+    return this.consumerApiService.triggerSignalEvent(identity, signalName, payload);
+  }
+
+}
