+++ src/kpi_service.ts
@@ -1,0 +1,54 @@
+import {IIdentity} from '@essential-projects/iam_contracts';
+
+import {IKpiApi} from '@process-engine/kpi_api_contracts';
+
+import {APIs, DataModels} from '@process-engine/management_api_contracts';
+
+export class KpiService implements APIs.IKpiManagementApi {
+
+  private readonly kpiApiService: IKpiApi;
+
+  constructor(kpiApiService: IKpiApi) {
+    this.kpiApiService = kpiApiService;
+  }
+
+  public async getRuntimeInformationForProcessModel(
+    identity: IIdentity,
+    processModelId: string,
+  ): Promise<Array<DataModels.Kpi.FlowNodeRuntimeInformation>> {
+
+    return this.kpiApiService.getRuntimeInformationForProcessModel(identity, processModelId);
+  }
+
+  public async getRuntimeInformationForFlowNode(
+    identity: IIdentity,
+    processModelId: string,
+    flowNodeId: string,
+  ): Promise<DataModels.Kpi.FlowNodeRuntimeInformation> {
+    return this.kpiApiService.getRuntimeInformationForFlowNode(identity, processModelId, flowNodeId);
+  }
+
+  public async getActiveTokensForProcessModel(identity: IIdentity, processModelId: string): Promise<Array<DataModels.Kpi.ActiveToken>> {
+    return this.kpiApiService.getActiveTokensForProcessModel(identity, processModelId);
+  }
+
+  public async getActiveTokensForCorrelationAndProcessModel(
+    identity: IIdentity,
+    correlationId: string,
+    processModelId: string,
+  ): Promise<Array<DataModels.Kpi.ActiveToken>> {
+    return this.kpiApiService.getActiveTokensForCorrelationAndProcessModel(identity, correlationId, processModelId);
+  }
+
+  public async getActiveTokensForProcessInstance(
+    identity: IIdentity,
+    processInstanceId: string,
+  ): Promise<Array<DataModels.Kpi.ActiveToken>> {
+    return this.kpiApiService.getActiveTokensForProcessInstance(identity, processInstanceId);
+  }
+
+  public async getActiveTokensForFlowNode(identity: IIdentity, flowNodeId: string): Promise<Array<DataModels.Kpi.ActiveToken>> {
+    return this.kpiApiService.getActiveTokensForFlowNode(identity, flowNodeId);
+  }
+
+}
