+++ src/user_task_service.ts
@@ -1,0 +1,78 @@
+import {Subscription} from '@essential-projects/event_aggregator_contracts';
+import {IIdentity} from '@essential-projects/iam_contracts';
+
+import {IConsumerApi} from '@process-engine/consumer_api_contracts';
+import {APIs, DataModels, Messages} from '@process-engine/management_api_contracts';
+
+export class UserTaskService implements APIs.IUserTaskManagementApi {
+
+  private readonly consumerApiService: IConsumerApi;
+
+  constructor(consumerApiService: IConsumerApi) {
+    this.consumerApiService = consumerApiService;
+  }
+
+  public async getUserTasksForProcessModel(identity: IIdentity, processModelId: string): Promise<DataModels.UserTasks.UserTaskList> {
+    return this.consumerApiService.getUserTasksForProcessModel(identity, processModelId);
+  }
+
+  public async getUserTasksForProcessInstance(identity: IIdentity, processInstanceId: string): Promise<DataModels.UserTasks.UserTaskList> {
+    return this.consumerApiService.getUserTasksForProcessInstance(identity, processInstanceId);
+  }
+
+  public async getUserTasksForCorrelation(identity: IIdentity, correlationId: string): Promise<DataModels.UserTasks.UserTaskList> {
+    return this.consumerApiService.getUserTasksForCorrelation(identity, correlationId);
+  }
+
+  public async getUserTasksForProcessModelInCorrelation(
+    identity: IIdentity,
+    processModelId: string,
+    correlationId: string,
+  ): Promise<DataModels.UserTasks.UserTaskList> {
+    return this.consumerApiService.getUserTasksForProcessModelInCorrelation(identity, processModelId, correlationId);
+  }
+
+  public async finishUserTask(
+    identity: IIdentity,
+    processInstanceId: string,
+    correlationId: string,
+    userTaskInstanceId: string,
+    userTaskResult: DataModels.UserTasks.UserTaskResult,
+  ): Promise<void> {
+    return this.consumerApiService.finishUserTask(identity, processInstanceId, correlationId, userTaskInstanceId, userTaskResult);
+  }
+
+  // Notifications
+  public async onUserTaskWaiting(
+    identity: IIdentity,
+    callback: Messages.CallbackTypes.OnUserTaskWaitingCallback,
+    subscribeOnce?: boolean,
+  ): Promise<Subscription> {
+    return this.consumerApiService.onUserTaskWaiting(identity, callback, subscribeOnce);
+  }
+
+  public async onUserTaskFinished(
+    identity: IIdentity,
+    callback: Messages.CallbackTypes.OnUserTaskFinishedCallback,
+    subscribeOnce?: boolean,
+  ): Promise<Subscription> {
+    return this.consumerApiService.onUserTaskFinished(identity, callback, subscribeOnce);
+  }
+
+  public async onUserTaskForIdentityWaiting(
+    identity: IIdentity,
+    callback: Messages.CallbackTypes.OnUserTaskWaitingCallback,
+    subscribeOnce?: boolean,
+  ): Promise<Subscription> {
+    return this.consumerApiService.onUserTaskForIdentityWaiting(identity, callback, subscribeOnce);
+  }
+
+  public async onUserTaskForIdentityFinished(
+    identity: IIdentity,
+    callback: Messages.CallbackTypes.OnUserTaskFinishedCallback,
+    subscribeOnce?: boolean,
+  ): Promise<Subscription> {
+    return this.consumerApiService.onUserTaskForIdentityFinished(identity, callback, subscribeOnce);
+  }
+
+}
